chore: [#185438439] split profile tests
